validations producto expirado

+++ CRUDS.sql
@@ -1498,8 +1498,13 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
         
 		IF (pOperacion = 'DELETE') THEN
 			IF ((SELECT COUNT(*) FROM BODEGA_SUCURSAL_PRODUCTO WHERE Cod_Bode_Sucu_Produ = pCodBodega) > 0) THEN
-				DELETE FROM BODEGA_SUCURSAL_PRODUCTO
-				WHERE Cod_Bode_Sucu_Produ = pCodBodega;
+				IF ((SELECT COUNT(*) FROM PRODUCTO_EXPIRADO WHERE Cod_Bode_Sucu_Produ = pCodBodega) = 0) THEN
+					DELETE FROM BODEGA_SUCURSAL_PRODUCTO
+					WHERE Cod_Bode_Sucu_Produ = pCodBodega;
+				ELSE
+					SET msgError = 'Imposible eliminar, código de bodega está asociado a producto expirado';
+					SELECT msgError;
+				END IF;
             ELSE
 				SET msgError = 'No se puede eliminar, el código de bodega de sucursal no existe';
 				SELECT msgError;
@@ -1888,8 +1893,8 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
 			END IF;
 		END IF;
 	
-    #el codigo de bono no es solicitado al crear el bono, se así se evita el mensaje
-	ELSEIF ((pCodEmpleado IS NOT NULL) AND (pOperacio!='CREATE')) THEN
+    #el codigo de bono no es solicitado al crear el bono, así se evita el mensaje
+	ELSEIF ((pCodEmpleado IS NOT NULL) AND (pOperacion!='CREATE')) THEN
 		SET msgError = 'El codigo de bono es vacío';
         SELECT msgError;
 	END IF;
@@ -2156,27 +2161,73 @@ END;
 
 /*------------------------------------------------------------ PRODUCTO_EXPIRADO ------------------------------------------------------------*/
 delimiter //
-CREATE PROCEDURE CRUD_PRODUCTO_EXPIRADO(pCodProdu INT, pNombProdu VARCHAR(50), pFecha DATE, pOperacion VARCHAR(10)) 
+CREATE PROCEDURE CRUD_PRODUCTO_EXPIRADO(pCodProduExp INT, pCodProdu INT, pCantProExp INT, pBodeSucuProdu INT, pOperacion VARCHAR(10)) 
 BEGIN
-	IF (pOperacion = 'CREATE') THEN
-		INSERT INTO PRODUCTO_EXPIRADO (Cod_Producto,Nombre_Producto) 
-        VALUES (pCodProdu,pNombProdu,pFecha);
-    END IF;
+DECLARE msgError VARCHAR(70) DEFAULT '';
     
-    IF (pOperacion = 'READ') THEN
-		SELECT *
-        FROM PRODUCTO_EXPIRADO
-        WHERE pCodProdu = PRODUCTO_EXPIRADO.Cod_Producto;
-    END IF;
-    /*
-    IF (pOperacion = 'UPDATE')  THEN
-		UPDATE PRODUCTO_EXPIRADO
-        SET 
-    END IF;
-    */
-    IF (pOperacion = 'DELETE') THEN
-		DELETE FROM PRODUCTO_EXPIRADO
-        WHERE pCodProdu = PRODUCTO_EXPIRADO.Cod_Producto;
-    END IF;
-END
+	IF (pOperacion = 'CREATE') THEN
+		IF ((SELECT COUNT(*) FROM PRODUCTO WHERE Cod_Producto = pCodProdu) = 0) THEN
+				IF ((SELECT COUNT(*) FROM BODEGA_SUCURSAL_PRODUCTO WHERE Cod_Bode_Sucu_Produ = pBodeSucuProdu) > 0) THEN
+					IF (pCantProExp IS NOT NULL AND pCantProExp>=0) THEN
+							INSERT INTO PRODUCTO_EXPIRADO (Cod_Producto, Cantidad_Productos, Cod_Bode_Sucu_Produ) 
+							VALUES (pCodProdu, pCantProExp, pBodeSucuProdu);
+					ELSE
+						SET msgError = 'La cantidad de productos expirados no es válida';
+						SELECT msgError;
+					END IF;
+				ELSE
+					SET msgError = 'El código de bodega sucursal no existe';
+					SELECT msgError;
+				END IF;
+		ELSE
+			SET msgError = 'El código de producto no existe';
+			SELECT msgError;
+		END IF;
+	END IF;
+    
+	IF (pCodProdu IS NOT NULL) THEN
+    
+		IF (pOperacion = 'READ') THEN
+			IF ((SELECT COUNT(*) FROM PRODUCTO_EXPIRADO WHERE Cod_Producto_Exp = pCodProduExp) > 0) THEN
+				SELECT Cod_Producto, Cantidad_Productos, Cod_Bode_Sucu_Produ
+				FROM PRODUCTO_EXPIRADO
+				WHERE Cod_Producto_Exp = pCodProduExp;
+			ELSE
+				SET msgError = 'El codigo producto expirado no existe, imposible realizar la busqueda';
+				SELECT msgError;
+			END IF;
+		END IF;
+
+		IF (pOperacion = 'UPDATE')  THEN
+			IF ((SELECT COUNT(*) FROM PRODUCTO_EXPIRADO WHERE Cod_Producto_Exp = pCodProduExp) > 0) THEN
+				IF (pCantProExp IS NOT NULL AND pCantProExp>=0) THEN
+						UPDATE PRODUCTO_EXPIRADO
+						SET Cantidad_Productos=IFNULL(pCantProExp,Cantidad_Productos)
+						WHERE Cod_Producto_Exp = pCodProduExp;
+					ELSE
+						SET msgError = 'Cantidad de productos expirados es inválida,imposible actualizar datos';
+						SELECT msgError;
+					END IF;
+            ELSE
+				SET msgError = 'El codigo de producto expirado no existe, no se puede actualizar datos';
+				SELECT msgError;
+			END IF;
+		END IF;
+        
+		IF (pOperacion = 'DELETE') THEN
+			IF ((SELECT COUNT(*) FROM PRODUCTO_EXPIRADO WHERE Cod_Producto_Exp = pCodProduExp) > 0) THEN
+				DELETE FROM PRODUCTO_EXPIRADO
+				WHERE Cod_Producto_Exp = pCodProduExp;
+			 ELSE
+				SET msgError = 'El codigo de producto expirado no existe, no se puede eliminar datos';
+				SELECT msgError;
+			END IF;
+		END IF;
+        
+    #el codigo de producto expirado no es solicitado al crear el registro, así se evita el mensaje
+	ELSEIF ((pCodEmpleado IS NOT NULL) AND (pOperacion!='CREATE')) THEN
+		SET msgError = 'El codigo de producto expirado es vacío';
+        SELECT msgError;
+	END IF;
+END;
 //
